+++ R/enrich.R
@@ -50,10 +50,13 @@ setMethod("enrich", "partition_bundle", function(.Object, mc = FALSE, progress =
 
 
 #' @rdname kwic-class
-setMethod("enrich", "kwic", function(.Object, meta = NULL, table = FALSE){
-  if (length(meta) > 0){
+setMethod("enrich", "kwic", function(.Object, s_attributes = NULL, table = FALSE, ...){
+  
+  if ("meta" %in% names(list(...))) s_attributes <- list(...)[["meta"]]
+  
+  if (length(s_attributes) > 0L){
     metainformation <- lapply(
-      meta,
+      s_attributes,
       function(metadat){
         cposToGet <- .Object@cpos[which(.Object@cpos[["position"]] == 0)][, .SD[1], by = "hit_no", with = TRUE][["cpos"]]
         strucs <- CQI$cpos2struc(.Object@corpus, metadat, cposToGet)
@@ -65,9 +68,9 @@ setMethod("enrich", "kwic", function(.Object, meta = NULL, table = FALSE){
       }
     )
     metainformation <- data.frame(metainformation, stringsAsFactors = FALSE)
-    colnames(metainformation) <- meta
+    colnames(metainformation) <- s_attributes
     .Object@table <- data.frame(metainformation, .Object@table)
-    .Object@metadata <- c(meta, .Object@metadata)
+    .Object@metadata <- c(s_attributes, .Object@metadata)
   }
   
   if (table){